feat: InputText 컴포넌트 수정
