+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import CommonBadge from '@/shared/common/ui/badge';
 import Chip from '@/shared/common/ui/chip';
 import CommonButton from '@/shared/common/ui/common-button';
 import CommonInput from '@/shared/common/ui/common-input';
-import InputText from '@/shared/common/ui/InputText';
+import InputText from '@/shared/common/ui/input-text';
 import SelectBox from '@/shared/common/ui/select/index';
 import TagClock from '@/shared/common/ui/tag-clock';
 import { selectBoxMenuItems } from '@/shared/fixture/selectbox-menu-items';
@@ -16,11 +16,11 @@ export default function Home() {
         placeholder="장소를 선택해주세요"
         menuItems={selectBoxMenuItems}
       />
+      <InputText placeholder="남겨주신 리뷰는 프로그램 운영 및 다른 회원 분들께 큰 도움이 됩니다." />
       <CommonButton variant="outlined" size="lg">
         생성하기
       </CommonButton>
       <Chip selected={true}>전체</Chip>
-      <InputText placeholder="내용을 입력해주세요" />
       <CommonInput />
       <TagClock message="오늘 21시 마감" variant="default" />
     </div>
